Separate code into different files
And rename most examples of "prior" to "priors" to avoid conflict with prior.py
