+++ tests/test_inference.py
@@ -31,7 +31,7 @@ import tsinfer
 
 import tsdate
 from tests import utility_functions
-from tsdate.date import LOG, LIN
+from tsdate.base import LOG, LIN
 
 
 class TestPrebuilt(unittest.TestCase):
@@ -168,11 +168,11 @@ class TestSimulated(unittest.TestCase):
         ts = msprime.simulate(
             sample_size=10, length=2e6, Ne=10000, mutation_rate=1e-8,
             recombination_rate=1e-8, random_seed=11)
-        prior = tsdate.build_prior_grid(ts, timepoints=10, approximate_prior=None)
-        dated_ts = tsdate.date(ts, Ne=10000, mutation_rate=1e-8, prior=prior,
-                               probability_space="linear")
-        maximized_ts = tsdate.date(ts, Ne=10000, mutation_rate=1e-8, prior=prior,
-                                   method='maximization', probability_space="linear")
+        priors = tsdate.build_prior_grid(ts, timepoints=10, approximate_priors=None)
+        dated_ts = tsdate.date(ts, Ne=10000, mutation_rate=1e-8, priors=priors,
+                               probability_space=LIN)
+        maximized_ts = tsdate.date(ts, Ne=10000, mutation_rate=1e-8, priors=priors,
+                                   method='maximization', probability_space=LIN)
         self.ts_equal_except_times(ts, dated_ts)
         self.ts_equal_except_times(ts, maximized_ts)
 
@@ -196,11 +196,11 @@ class TestSimulated(unittest.TestCase):
                 if not internal_edge_removed:
                     continue
             tables.edges.add_row(*row)
-        multi_root_ts = tables.tree_sequence()
-        good_prior = tsdate.build_prior_grid(ts)
-        self.assertRaises(ValueError, tsdate.build_prior_grid, multi_root_ts)
-        self.assertRaises(ValueError, tsdate.date, multi_root_ts, 1, 2)
-        self.assertRaises(ValueError, tsdate.date, multi_root_ts, 1, 2, None, good_prior)
+        multiroot_ts = tables.tree_sequence()
+        good_priors = tsdate.build_prior_grid(ts)
+        self.assertRaises(ValueError, tsdate.build_prior_grid, multiroot_ts)
+        self.assertRaises(ValueError, tsdate.date, multiroot_ts, 1, 2)
+        self.assertRaises(ValueError, tsdate.date, multiroot_ts, 1, 2, None, good_priors)
 
     def test_non_contemporaneous(self):
         samples = [
