+++ tsdate/util.py
@@ -9,8 +9,8 @@
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
 #
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
 #
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -19,9 +19,26 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+"""
+Utility functions for tsdate. Many of these can be removed when tskit is updated to
+a more recent version which has the functionality built-in
+"""
 
 
-from .date import date # NOQA
-from .prior import build_grid as build_prior_grid # NOQA
-from .provenance import __version__ # NOQA
-from .cache import * # NOQA
+def edge_span(edge):
+    return edge.right - edge.left
+
+
+def tree_num_children(tree, node):
+    return len(tree.children(node))
+
+
+def get_single_root(tree):
+    # TODO - use new 'root_threshold=2' to avoid having to check isolated nodes
+    topological_roots = [r for r in tree.roots if tree_num_children(tree, r) != 0]
+    if len(topological_roots) > 1:
+        raise ValueError(
+            "Invalid tree sequence: tree {} has >1 root".format(tree.index))
+    if len(topological_roots) == 0:
+        return None  # Empty tree
+    return topological_roots[0]
